Added same changes to wsGen
